Match placements against some criteria
So far, we have a couple of fields we are matching against,
and are treated as optional, but preferred. We also use a
function score to weight the distance, so closer premises are
preferred.

I’ve also added the `class-transformer` package so we can
easily cast query objects to typescript classes.