Index on premises, instead of beds
We should be treating premises as first class objects, with the
beds and bookings as second class objects. Before we were
indexing a lot of data with not much use, as most of the things
that relate to beds actually relate to the premises.

I’ve also discovered there is a `range` data type, so I’m indexing
the bookings as ranges, which will allow us to do intersection
queries