Document 'alpha' bed-matching using OpenSearch
We've demonstrated the ability to independently score these dimensions of:

- room / AP characteristic
- availability
- distance

We anticipate that the Beta team will want to build an extensive test suite
which allows the "matching algorithm" to be refined against:

- a realistic set of bookings and
- a varied corpus of queries and expected rankings

In this way the effectiveness of the matcher can be systematically improved.
As deficiencies are uncovered new test examples can be added to the suite.

In these notes we provide an overview to the techniques used with OpenSearch
as a primer for the Beta team taking this forward.