Refactor the form to use classes
This ensures we can make tests simpler and more [solitary](https://martinfowler.com/bliki/UnitTest.html),
allowing us to test each individual step in isolation.

Steps now live in a folder called `steps` and inherit from a `Step`
class. Each step class requires us to set a `nextStep` method
and a `dto` method. The `valid` method then validates the dto,
and sets the errors on the inheriting class, also converting the errors
to a usable state (something that was handled in the form before).

We then export a `stepList` in the steps index file, which maps each
step to a class (we have to define steps we haven’t yet built out to
return a bare `Step` class, so we can be sure they show up as an
allowed step name)

Then, in the form class, we accept the step name and the params
as before, the `getStep` class then picks the step class from the
`stepList` object, and we can then check the validity and the next
place to go from there.