Pass a strongly typed request into the form
Express uses [Generics](https://www.digitalocean.com/community/tutorials/typescript-generics-in-typescript)
to allow us to specify the shape of the request body we’re
getting in our route. As everything we’re passing to the form
is gleaned from the request, we can use some of this to strongly
type the data we expect from our route. I’ve also used the
`ReferralApplicationBody` type to strongly type the session
data that’s stored too.