Add a Question method
This accepts a step and a question name, which then reads a JSON
file to get a nunjucks widget name and the options to pass to it.

The `present()` method then adds the error message (if present)
to the arguments and then also reads the body data and adds
`checked` or `value` arguments to the arguments, based on the
data passed to the step. It then renders the data to a HTML string
using Nunjucks.

I could have passed the widget name and the arguments directly
to the template, but doing it this way allows us to treat any
conditional questions as first-class too, rendering them in the
`conditional` attribute of a radio button’s value.