Fetch information about a section from a JSON file
This allows us to specify things like the next and previous
sections in JSON, in a similar way to how we handle
steps. This can also be extended in future to add logic.

I’ve had to refactor the form spec a bit too, to make sure
we now mock the sections, rather than use real life
objects