Read session from saved json file
‘form’s initialise method is either called with a request whose ‘body’
is the response to a question in the UI or a request with an empty
‘body’.

Here if the body doesn’t have any properties we attempt to read the
session from the saved json file we created earlier. To prevent
crashes we have to check the file exists before reading it. If the
‘body’ has no properties AND the saved session doesn’t exist we return
an empty object