Write data to json when persisting
here we write the session to a json file with the sessionID as a file
name when the ‘persistData’ method is called. In the beta stage this
will probably be replaced with a write to the db but this proves the
concept for now.

I haven’t been able to work out why but sometimes ‘saveSession’ is
called with a function as the first argument. this call isn’t showing
up when logging or debugging but it does happen. For now to prevent it
 I have added a check that the ‘username’ is a string but these
erroneous calls should be investigated before/during beta.